+++ com/pdfjet/Page.java
@@ -1712,9 +1712,9 @@ final public class Page {
                 list.add(new Point(p.x - p.r, p.y + p.r));
                 list.add(new Point(p.x - p.r, p.y - p.r));
                 drawPath(list, p.getPathOperator());
-            } else if (p.shape == Shape.H_DASH) {
+            } else if (p.shape == Shape.HORIZONTAL_DASH) {
                 drawLine(p.x - p.r, p.y, p.x + p.r, p.y);
-            } else if (p.shape == Shape.V_DASH) {
+            } else if (p.shape == Shape.VERTICAL_DASH) {
                 drawLine(p.x, p.y - p.r, p.x, p.y + p.r);
             } else if (p.shape == Shape.X_MARK) {
                 drawLine(p.x - p.r, p.y - p.r, p.x + p.r, p.y + p.r);